1.完成了“当天订货单列表接口，及订货单提交接口”，传递参数json list，如：[{"goodsCode":"111111","price":"2.12","goodTotal":"10"},{"goodsCode":"222222","price":"3","goodTotal":"15"}]
2.修正了框架 事务管理

+++ supershopowner/src/main/java/com/xiaomai/supershopowner/entity/Order.java
@@ -13,7 +13,7 @@ public class Order {
 
     private Integer orderNumber;
 
-    private Long totalValue;
+    private Double totalValue;
 
     public Integer getId() {
         return id;
@@ -55,11 +55,12 @@ public class Order {
         this.orderNumber = orderNumber;
     }
 
-    public Long getTotalValue() {
-        return totalValue;
-    }
+	public Double getTotalValue() {
+		return totalValue;
+	}
 
-    public void setTotalValue(Long totalValue) {
-        this.totalValue = totalValue;
-    }
+	public void setTotalValue(Double totalValue) {
+		this.totalValue = totalValue;
+	}
+    
 }
